fix: [Common] Fix Coverity issues (#1936)
Typecasting integer variable to prevent overflow (CWE 190)

+++ BootloaderCommonPkg/Library/MmcAccessLib/SdMmcPciHci.c
@@ -644,8 +644,8 @@ SdMmcHcClockSupply (
     return EFI_INVALID_PARAMETER;
   }
 
-  if (ClockFreq > (BaseClkFreq * 1000)) {
-    ClockFreq = BaseClkFreq * 1000;
+  if (ClockFreq > ((UINT64)BaseClkFreq * 1000)) {
+    ClockFreq = (UINT64)BaseClkFreq * 1000;
   }
 
   //
